refactor: move text overlay configures to text.go
